# docs/api/openapi/cordum-rest.yaml
openapi: 3.0.3
info:
  title: Cordum REST API
  version: 1.0.0
  description: |
    Safety-first agent orchestration platform. Cordum provides a control plane
    for submitting, scheduling, and governing AI agent jobs with built-in safety
    policies, approval workflows, dead-letter queues, and audit logging.

    ## Authentication
    All protected endpoints require either an API key (`X-API-Key` header) or a
    Bearer JWT token obtained from the login endpoint. Public endpoints are
    explicitly marked and require no credentials.

    ## Multi-Tenancy
    Protected routes require the `X-Tenant-ID` header to scope requests to the
    correct tenant partition.
  contact:
    name: Cordum Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8081
    description: Development
...
security:
  - apiKey: []
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/workflows/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
      - $ref: "#/components/parameters/TenantID"
    get:
      operationId: getWorkflow
      summary: Get a workflow definition
      tags: [Workflows]
      responses:
        "200":
          description: Workflow definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteWorkflow
      summary: Delete a workflow definition
      tags: [Workflows]
      responses:
        "204":
          description: Workflow deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/locks/release:
    post:
      operationId: releaseLock
      summary: Release a distributed lock
      tags: [Locks]
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockRequest"
      responses:
        "200":
          description: Lock released
          content:
            application/json:
              schema:
                type: object
                properties:
                  lock:
                    $ref: "#/components/schemas/Lock"
                  released:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued by the Cordum admin.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login.

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant isolation header (required on all protected routes).
      schema:
        type: string
...
    WorkflowID:
      name: id
      in: path
      required: true
      description: Workflow ID
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Unauthorized — missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
      required: [error, status]
...
    WorkflowDefinition:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        team_id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        timeout_sec:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        config:
          type: object
          additionalProperties: true
      required: [id, name, steps]
    WorkflowStep:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          description: Step type (job, switch, parallel, loop, transform, storage, subworkflow, approval)
        name:
          type: string
        depends_on:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        timeout_sec:
          type: integer
        retry:
          type: object
          properties:
            max_attempts:
              type: integer
            backoff_sec:
              type: integer
      required: [id, type]
...
    Lock:
      type: object
      properties:
        resource:
          type: string
        mode:
          type: string
          enum: [EXCLUSIVE, SHARED]
        owners:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    LockRequest:
      type: object
      properties:
        resource:
          type: string
        owner:
          type: string
        mode:
          type: string
          enum: [EXCLUSIVE, SHARED]
        ttl_ms:
          type: integer
          description: Lock time-to-live in milliseconds
      required: [resource, owner]

    # --- Packs ------------------------------------------------------------
...